replaces instrumentid alias via int64.

+++ examples/demo/metronome_demo.cpp
@@ -67,7 +67,7 @@ int main(int /*argc*/, char* /*argv*/[]) {
         kRelease}}));
 
   // Beat callback.
-  const auto beat_callback = [&](double, int beat) {
+  const auto beat_callback = [&](int64, int beat) {
     const int current_bar = beat / kNumBeats;
     const int current_beat = beat % kNumBeats;
     LOG(INFO) << "Tick " << current_bar << "." << current_beat;
